🏗️ Change the architecture for zod schemas
