+++ features/authentication/update-password-form.tsx
@@ -3,7 +3,7 @@ import { Alert, Button, Group, PasswordInput, Stack } from "@mantine/core";
 import { IconInfoCircle } from "@tabler/icons-react";
 import { useActionState, useMemo } from "react";
 import { updatePasswordAction } from "@/lib/actions/authentication";
-import type { UpdatePasswordInput } from "@/zod-schemas/authentication";
+import type { UpdatePasswordInput } from "@/lib/schemas/authentication.schemas";
 
 type UpdatePasswordInputKeys = keyof UpdatePasswordInput;
