+++ theme/index.ts
@@ -7,69 +7,32 @@ const figtree = Figtree({
 	display: "swap",
 });
 
-// Generating 10-shade palettes for our accent colors
-const focusBlue: MantineColorsTuple = [
-	"#eaf0f8",
-	"#d0dceb",
-	"#b3c6de",
-	"#95b1d2",
-	"#7c9dc7",
-	"#6a8fc1",
-	"#5d88be",
-	"#4a75a9",
-	"#416896",
-	"#355a83",
-];
-const growthGreen: MantineColorsTuple = [
-	"#edf2ee",
-	"#d9e2da",
-	"#c1cfc3",
-	"#a8bca9",
-	"#92aa93",
-	"#839f84",
-	"#78977a",
-	"#68846b",
-	"#5d755f",
-	"#506652",
-];
-const inspirationPeach: MantineColorsTuple = [
-	"#faede7",
-	"#f1d8cc",
-	"#e7c0ae",
-	"#dea78e",
-	"#d79376",
-	"#d48c70",
-	"#cf8366",
-	"#b77156",
-	"#a4654d",
-	"#915742",
-];
-const wisdomViolet: MantineColorsTuple = [
-	"#f2eff4",
-	"#e2dbe5",
-	"#cfc3d4",
-	"#bbabcc",
-	"#a996c5",
-	"#9d8ac0",
-	"#9681bc",
-	"#836fA7",
-	"#756394",
-	"#675681",
-];
-const clarityGold: MantineColorsTuple = [
-	"#f9f3e9",
-	"#efeadb",
-	"#e4dcc2",
-	"#dacfa8",
-	"#d2c390",
-	"#cdae78",
-	"#cbac6c",
-	"#b3975a",
-	"#a0874e",
-	"#8d7641",
+const mutedSalmon: MantineColorsTuple = [
+	"#fdf5f3",
+	"#f9e9e6",
+	"#f1d6cd",
+	"#eac2b3",
+	"#e2ae9d",
+	"#dca18e",
+	"#d99b86", // The closest shade to the target #E58B78
+	"#c08573",
+	"#aa7565",
+	"#946456",
 ];
 
-// Semantic Colors
+const cozyGray: MantineColorsTuple = [
+	"#FAF9F7", // 0: Linen White (Primary Background)
+	"#F4F3F1", // 1: Parchment (Secondary Background / Cards)
+	"#EAE9E7", // 2: Subtle Line (Borders)
+	"#DDDBD8", // 3: UI Element Hover
+	"#C7C6C3", // 4: Disabled Elements
+	"#A8A7A4", // 5: Medium Gray
+	"#8A8986", // 6: Muted UI Elements / Secondary Text
+	"#646361", // 7:
+	"#4D4C4A", // 8:
+	"#3F3D3C", // 9: Charcoal (Primary Text)
+];
+
 const successGreen: MantineColorsTuple = [
 	"#ebf7f0",
 	"#d3eadd",
@@ -107,36 +70,17 @@ const errorRed: MantineColorsTuple = [
 	"#8f3e3e",
 ];
 
-// Our custom neutral palette based on the original design
-const brandGray: MantineColorsTuple = [
-	"#EFF0F3", // 0: Paper White (Background)
-	"#E7E8EC", // 1: Light Slate (UI Elements)
-	"#E0E1E6", // 2: Subtle Line (Borders)
-	"#D8D9E0", // 3
-	"#CDCED7", // 4
-	"#B9BBC6", // 5
-	"#8B8D98", // 6: Medium Gray (Secondary Text/Icons)
-	"#80828D", // 7
-	"#62636C", // 8: Charcoal (Primary Text)
-	"#1E1F24", // 9
-];
-
 export const theme = createTheme({
 	/* ### Base Properties ### */
 	fontFamily: `${figtree.style.fontFamily}, Verdana, sans-serif`,
-	primaryColor: "focusBlue",
+	primaryColor: "mutedSalmon",
 	primaryShade: 6,
 
 	/* ### Color Definitions ### */
 	colors: {
 		// Brand colors
-		focusBlue,
-		growthGreen,
-		inspirationPeach,
-		wisdomViolet,
-		clarityGold,
-		// Custom gray scale
-		brandGray,
+		mutedSalmon,
+		cozyGray,
 		// Semantic colors
 		successGreen,
 		warningAmber,
@@ -144,31 +88,33 @@ export const theme = createTheme({
 	},
 
 	/* ### Base Styles ### */
-	white: "#fff",
-	black: "#1E1F24",
+	// Override white/black to align with our cozy palette
+	white: cozyGray[0], // Linen White
+	black: cozyGray[9], // Charcoal
 
 	/* ### Custom Properties for Easy Access ### */
 	other: {
-		paperWhite: brandGray[0],
-		lightSlate: brandGray[1],
-		charcoalText: brandGray[9],
+		linenWhite: cozyGray[0],
+		parchment: cozyGray[1],
+		charcoalText: cozyGray[9],
+		stoneGray: cozyGray[6],
 	},
 
 	/* ### Default Component Overrides ### */
 	components: {
 		Paper: {
 			defaultProps: {
-				bg: "brandGray.0", // Use Paper White for Paper components
+				bg: "cozyGray.0",
 			},
 		},
 		Card: {
 			defaultProps: {
-				bg: "brandGray.1", // Use Light Slate for Card components
+				bg: "cozyGray.1",
 			},
 		},
 		AppShell: {
 			defaultProps: {
-				bg: "brandGray.0",
+				bg: "cozyGray.0",
 			},
 		},
 		Button: {
@@ -178,12 +124,12 @@ export const theme = createTheme({
 		},
 		Text: {
 			defaultProps: {
-				c: "brandGray.9", // Default text color is Charcoal
+				c: "cozyGray.9",
 			},
 		},
 		Title: {
 			defaultProps: {
-				c: "brandGray.9",
+				c: "cozyGray.9",
 			},
 		},
 	},
